Compress session log and resident drain tests
